test/fixtures/wpt/FileAPI/reading-data-section/FileReader-event-handler-attributes.html

<!doctype html>
<meta charset="utf-8">
<title>FileReader event handler attributes</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
var attributes = [
  "onloadstart",
  "onprogress",
  "onload",
  "onabort",
  "onerror",
  "onloadend",
];
attributes.forEach(function(a) {
  test(function() {
    var reader = new FileReader();
    assert_equals(reader[a], null,
                  "event handler attribute should initially be null");
  }, "FileReader." + a + ": initial value");
});
</script>
